backend/requirements.txt
Django~=5.2
asgiref
django-ckeditor-5
django-ckeditor
requests
django-environ
psycopg2-binary
pillow
python-dateutil
icalendar
django-ical
django-admin-ordering==0.16.1
whitenoise
channels
channels_redis
websocket_client
six
instaloader
schedule
boto3
django-storages
django-tables2
django-filter
django-bootstrap3
django_cleanup
daphne
celery[redis]
gunicorn
gspread
djangorestframework
